+++ fvm/src/executor/default.rs
@@ -83,6 +83,7 @@ where
                 machine,
                 msg.gas_limit,
                 sender_id,
+                msg.from,
                 msg.sequence,
                 msg.gas_premium.clone(),
             );