+++ fvm/tests/dummy.rs
@@ -214,6 +214,7 @@ pub struct DummyCallManager {
     pub machine: DummyMachine,
     pub gas_tracker: GasTracker,
     pub origin: ActorID,
+    pub origin_address: Address,
     pub nonce: u64,
     pub test_data: Rc<RefCell<TestData>>,
     limits: DummyLimiter,
@@ -238,6 +239,7 @@ impl DummyCallManager {
                 nonce: 0,
                 test_data: rc,
                 limits: DummyLimiter::default(),
+                origin_address: Address::new_id(0),
             },
             cell_ref,
         )
@@ -256,6 +258,7 @@ impl DummyCallManager {
                 nonce: 0,
                 test_data: rc,
                 limits: DummyLimiter::default(),
+                origin_address: Address::new_id(0),
             },
             cell_ref,
         )
@@ -269,6 +272,7 @@ impl CallManager for DummyCallManager {
         machine: Self::Machine,
         _gas_limit: i64,
         origin: ActorID,
+        origin_address: Address,
         nonce: u64,
         gas_premium: TokenAmount,
     ) -> Self {
@@ -280,6 +284,7 @@ impl CallManager for DummyCallManager {
             machine,
             gas_tracker: GasTracker::new(Gas::new(i64::MAX), Gas::new(0), gas_premium),
             origin,
+            origin_address,
             nonce,
             test_data: rc,
             limits,
@@ -350,7 +355,16 @@ impl CallManager for DummyCallManager {
         self.nonce
     }
 
-    fn next_actor_idx(&mut self) -> u64 {
+    fn next_actor_address(&self) -> Address {
+        todo!()
+    }
+
+    fn create_actor(
+        &mut self,
+        _code_id: Cid,
+        _actor_id: ActorID,
+        _predictable_address: Option<Address>,
+    ) -> kernel::Result<()> {
         todo!()
     }